회원가입 모듈화 적용
