+++ src/api/axiosApi/joinApi.js
@@ -1,0 +1,4 @@
+export const Join = {
+  join: "/api/registerBCrypt.do",
+  loginIdCheck: "/api/check_loginId.do",
+};
